Initial commit.
Depends on Drupal 8.2, at the time of commit requiring a core patch
https://www.drupal.org/node/2161337#comment-11452257, which is about to be
committed

This is still quite rough. Most prominently, it misses a UI to create repeat
rules. The basic architecture is solid, though. It uses a seperate table per
field to store the repeat occurences to not automatically always load them with
the entity. If there are many (think hundreds or thousands) occurences, this
should preserve sanity.

The formatter just displays all occurences at the moment and should be
considered a stub, as should be the widget.

Views works. Calendar needs to be patched.